Move send to clients method to NCToClient
