+++ Assets/Mirror/Runtime/NetworkConnection.cs
@@ -256,7 +256,7 @@ namespace Mirror
         //    would check max size and show errors internally. best to do it
         //    in one place in hlapi.
         // => it's important to log errors, so the user knows what went wrong.
-        static bool ValidatePacketSize(ArraySegment<byte> segment, int channelId)
+        protected internal static bool ValidatePacketSize(ArraySegment<byte> segment, int channelId)
         {
             if (segment.Count > Transport.activeTransport.GetMaxPacketSize(channelId))
             {
@@ -278,42 +278,7 @@ namespace Mirror
         // internal because no one except Mirror should send bytes directly to
         // the client. they would be detected as a message. send messages instead.
         List<int> singleConnectionId = new List<int>{-1};
-        internal virtual bool Send(ArraySegment<byte> segment, int channelId = Channels.DefaultReliable)
-        {
-            if (logNetworkMessages) Debug.Log("ConnectionSend " + this + " bytes:" + BitConverter.ToString(segment.Array, segment.Offset, segment.Count));
-
-            // validate packet size first.
-            if (ValidatePacketSize(segment, channelId))
-            {
-                // send to client or server
-                if (Transport.activeTransport.ClientConnected())
-                {
-                    return Transport.activeTransport.ClientSend(channelId, segment);
-                }
-                else if (Transport.activeTransport.ServerActive())
-                {
-                    singleConnectionId[0] = connectionId;
-                    return Transport.activeTransport.ServerSend(singleConnectionId, channelId, segment);
-                }
-            }
-            return false;
-        }
-
-        // Send to many. basically Transport.Send(connections) + checks.
-        internal static bool Send(List<int> connectionIds, ArraySegment<byte> segment, int channelId = Channels.DefaultReliable)
-        {
-            // validate packet size first.
-            if (ValidatePacketSize(segment, channelId))
-            {
-                // only the server sends to many, we don't have that function on
-                // a client.
-                if (Transport.activeTransport.ServerActive())
-                {
-                    return Transport.activeTransport.ServerSend(connectionIds, channelId, segment);
-                }
-            }
-            return false;
-        }
+        internal abstract bool Send(ArraySegment<byte> segment, int channelId = Channels.DefaultReliable);
 
         public override string ToString()
         {
